refactor(user): UserService → UserCommandService / UserQueryService 분리 [Story-4-4]
User BC application 레이어를 Command / Query로 분리. AC1·2·3·4 모두 충족.

신규:
- User/application/UserCommandService.java — @Service @Transactional. 쓰기 5 유스케이스:
  addUser / loginLocal / updateUser / deleteUser / socialLogin. 의존 5: PasswordEncoder /
  UserRepository / JwtService / TokenIssuer / SocialMemberRegistrar.
- User/application/UserQueryService.java — @Service @Transactional(readOnly=true)
  클래스 레벨. 조회 2 유스케이스: existUser / readUser. 의존 1: UserRepository.
  AC4 충족: dirty checking 정적 보장.

수정:
- User/presentation/UserController.java — 메서드별 주입 분기:
  - existUserApi / userMeApi → UserQueryService
  - loginLocal / joinApi / updateUserApi / deleteUserApi → UserCommandService
- User/presentation/SocialLoginController.java — UserService → UserCommandService 주입.
- User/application/SocialOAuthFlow.java — Javadoc 참조 정정 (UserService → UserCommandService).

삭제:
- User/application/UserService.java — 레거시 클래스 완전 제거 (AC3).

명세 변경 (사용자 합의 옵션 A):
- 메서드명 findOrRegisterSocialUser는 socialLogin 그대로 유지 (호출처 영향 최소화).
- CustomOAuth2UserService 위임은 Story-4-1에서 폐기되어 무관.
- ADR005 Command/Query record 적용은 본 Story 범위 외 — 별도 hygiene Story.

알려진 한계 (본 Story 외 처리):
- UserCommandService.socialLogin이 HttpServletResponse 직접 의존 — application 레이어
  servlet 누수. Controller로 토큰 발급 이관 후속.
- UserCommandService.deleteUser의 SecurityContextHolder 직접 호출 + AccessDeniedException
  → Story-5-3 범위.
- AccessDeniedException이 GlobalExceptionHandler 미처리 → 응답 형식 불일치. Story-5-3.
